% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modDemo.R
\name{add_Demo}
\alias{add_Demo}
\title{Merge in Demographic Data}
\usage{
add_Demo(dat, dat.columns, demo.list, demo.columns)
}
\arguments{
\item{dat}{Base data set, a filename (CSV, RData, RDS), or data.frame}

\item{dat.columns}{a named list that should specify columns in data.
\sQuote{id} is required. \sQuote{idvisit} may also be specified;
\sQuote{idvisit} can be used when there are multiple visits 
(i.e., several occasions) for the same subject.}

\item{demo.list}{demographic information, if available; the output from 
\code{\link{run_Demo}} or a correctly formatted data.frame}

\item{demo.columns}{a named list that should specify columns in demographic data;
\sQuote{id} is required. \sQuote{idvisit} may also be used to specify
columns for the unique idvisit. Any other columns
present in the demographic data are treated as covariates.}
}
\value{
data.frame
}
\description{
This module will merge demographic data with a base data set by id.
}
\examples{
user <- data.frame(id = c('A','A','B'), event = c('A.1','A.2','B.1'), enroll = '2025-02-25')
demo <- data.frame(id = c('A', 'B'), age = c(25, 45))
demo_w_visit <- data.frame(id = c('A','A','B'), event = c('A.1','A.2','B.1'), age = c(25, 26, 45))
add_Demo(user, list(id = 'id'), demo, list(id = 'id'))
add_Demo(user, list(id = 'id', idvisit = 'event'), demo_w_visit, list(id = 'id', idvisit = 'event'))
}
