% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modLabs.R
\name{add_Labs}
\alias{add_Labs}
\title{Merge in Lab Data}
\usage{
add_Labs(dat, dat.columns, lab.list, lab.columns, labPriorWindow = 7)
}
\arguments{
\item{dat}{Base data set, a filename (CSV, RData, RDS), or data.frame}

\item{dat.columns}{a named list that should specify columns in data.
\sQuote{id} and \sQuote{datetime} are required. \sQuote{datetime} can refer
to a single date-time variable (datetime = \sQuote{date_time}) or two
variables holding date and time separately (e.g., datetime = c(\sQuote{Date}, \sQuote{Time})).}

\item{lab.list}{lab data, if available; the output from \code{\link{run_Labs}} or 
a correctly formatted list}

\item{lab.columns}{a named list that should specify columns in lab data; \sQuote{id},
and \sQuote{datetime} are required. \sQuote{datetime} is the date and time when
the lab data was obtained, which can refer to a single date-time variable
(datetime = \sQuote{date_time}) or two variables holding date and time separately
(e.g., datetime = c(\sQuote{Date}, \sQuote{Time})). Any other columns present in lab
data are treated as lab values.}

\item{labPriorWindow}{This value sets the time frame window with the number of days
prior to the first date.time event in dat; defaults to 7.}
}
\value{
data.frame
}
\description{
This module will merge laboratory data with a base data set by closest date.
}
\examples{
user <- data.frame(id = c('A', 'B'), datetime = '2025-02-25 10:00')
labs <- structure(list(
    id = c("A", "A", "A", "B", "B", "B", "B", "B", "B", "B"),
    dt = structure(c(
      1738800000, 1756684800, 1775865600, 1726790400, 1737763200,
      1741910400, 1742169600, 1750118400, 1779062400, 1782345600
      ), class = c("POSIXct", "POSIXt"), tzone = ""),
    weight = c(155L, 163L, 169L, 170L, 172L, 175L, 183L, 187L, 197L, 198L),
    a1c = c(5.93, 5.97, 5.98, 5.99, 6.04, 6.04, 6.05, 6.08, 6.16, 6.17)
  ), class = "data.frame", row.names = c(NA, -10L)
)
add_Labs(user, list(id = 'id', datetime = 'datetime'), labs, list(id = 'id', datetime = 'dt'), labPriorWindow = 30)
}
