% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_quantile.R
\name{add_quantile}
\alias{add_quantile}
\title{Quantile-based discretization of column}
\usage{
add_quantile(col, n_quantile)
}
\arguments{
\item{col}{column of data.table to be discretized}

\item{n_quantile}{an integer specifying the number of quantiles}
}
\value{
vector of integer indicators of bins
}
\description{
Discretize column values into equal-sized bins
}
\examples{
\dontrun{
dem_dt <- copy(dem)
dem_dt[, monthly_income_quartile:=add_pctile(monthly_income, n_quantile=4)]

dt <- as.data.table(rep(10, 100))
dt[, decile:=add_pctile(V1, n_quantile=10)]
}
}
