% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastdoseUtils.R
\name{addLastDose}
\alias{addLastDose}
\title{Add Lastdose Data}
\usage{
addLastDose(buildData, lastdoseData)
}
\arguments{
\item{buildData}{data.frame, output of \code{build} function.}

\item{lastdoseData}{data.frame with columns filename, ld_start, lastdose,
raw_time, time_type}
}
\value{
a data.frame with the \sQuote{lastdose} column added.
}
\description{
Add lastdose data to data set
}
\details{
Lastdose is a datetime string associated with dose data.
}
