% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkDosingInterval.R
\name{build_lastdose}
\alias{build_lastdose}
\title{PK oral functions}
\usage{
build_lastdose(x, first_interval_hours = 336, ldCol = NULL)
}
\description{
code{build_lastdose}: build PK oral with lastdose information
}
\keyword{internal}
