% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkify.R
\name{chunkify}
\alias{chunkify}
\title{Chunkify a data.table into a list of data.tables.}
\usage{
chunkify(dt, chunksize)
}
\arguments{
\item{dt}{Name of data.table to be chunkfied [character]}

\item{chunk_num}{Rows per chunk [integer]}
}
\value{
TBC
}
\description{
Chunkify a data.table into a list of data.tables.
}
\examples{
dia_list <- chunkify(dia, 1000)
}
