% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkify_group.R
\name{chunkify_group}
\alias{chunkify_group}
\title{Chunkify a data.table into a list of data.tables preserving group-level clusters.}
\usage{
chunkify_group(dt, group_var, chunk_num)
}
\arguments{
\item{dt}{Name of data.table to be chunkfied [character]}

\item{group_var}{Name of clustering variable [character]}

\item{chunk_num}{Number of chunks to generate [integer]}
}
\value{
TBC
}
\description{
Chunkify a data.table into a list of data.tables preserving group-level clusters.
}
\examples{
dia_list <- chunkify_group(dia, group_var="empi", 10)
}
