% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.icd.list.R
\name{clean.icd.list}
\alias{clean.icd.list}
\title{Expand a list of icd9 code ranges.}
\usage{
clean.icd.list(list, icdcol, namescol, decimal = TRUE, out.name = namescol)
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Expand a list of icd9 code ranges.
}
\examples{
TBC
}
