% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_ci.R
\name{clustered_ci}
\alias{clustered_ci}
\title{Clustered confidence intervals}
\usage{
clustered_ci(data, obs_col_name, cluster_by_col_name, ci_level = 0.95)
}
\arguments{
\item{data}{data.table consisting, at minimum, of a column with the identifier}

\item{cluster_by_col_name}{column name by which to cluster. (character)}

\item{ci_level}{level of confidence for confidence intervals (provided as a decimal < 1). (numeric)}

\item{obs_col_name`}{name of the column in the data data.table that contains raw sample
values for which CI is desired (character)}
}
\value{
value that should be added and subtracted from the raw mean to yield the x% CIs
}
\description{
Calculate confidence intervals based on standard errors clustered by group. This was done using formula and explanations at
             http://www.fao.org/wairdocs/ILRI/x5436E/x5436e07.htm
             [NOTE formula SE = m/n*sqrt(1-f)..] in bold in section 5.2.1 has a mistake. The correct version is the one with numbers filled in below it.
}
\examples{
\dontrun{
data <- copy(ehR_cohort)
cis <- clustered_ci(data, obs = 'test', cluster_by_col_name = 'empi', ci_level = 0.95)
}
}
