% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_stat.R
\name{corr_stat}
\alias{corr_stat}
\title{Generate a correlation matrix, summary statistics and a distribution graph for the variables in a given dt.}
\usage{
corr_stat(dt, var_list_list, save = FALSE, path = NA)
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Generate a correlation matrix, summary statistics and a distribution graph for the variables in a given dt.
}
\examples{
TBC
}
