% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_dt.R
\name{cum_dt}
\alias{cum_dt}
\title{Count number of observations in a data.table falling into different bins of an identifier variable.}
\usage{
cum_dt(dt, id, var_, cutoff, external_count_dt = NA, split_var_list = NA,
  verbose = TRUE, ci_sd = FALSE, ci_type = "perc_prop",
  ci_clust = FALSE, ci_clust_var_name = NA, ci_level = 0.95)
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Count number of observations in a data.table falling into different bins of an identifier variable.
}
\examples{
TBC
}
