% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransformation.R
\name{dataTransformation}
\alias{dataTransformation}
\title{Data Transformation}
\usage{
dataTransformation(x, select, rename, modify)
}
\arguments{
\item{x}{a data.frame}

\item{select}{columns to select}

\item{rename}{character vector with names for all columns}

\item{modify}{list of expressions used to transform data set}
}
\value{
The modified data.frame
}
\description{
Convenience function for making small modifications to a data.frame.
}
