\docType{data}
\name{dd.baseline}
\alias{dd.baseline}
\title{dd.baseline}
\format{A data frame with 10000 observations on 1505 variables.}
\usage{
dd.baseline
}
\description{
Simulated baseline data example from a Phenome Wide Association Study (PheWAS) 
obtained at baseline before drug exposure. The dataset includes 1505 variables: 
subject identification number ('id'), drug exposure ('exposure'), 3 demographic 
variables ('age', 'race', 'gender'), and 1500 phenotypes.
}
\examples{
data(dataPheWAS)
}
\keyword{datasets}
