% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\docType{data}
\name{dd}
\alias{dd}
\title{dd}
\format{
A data frame with 10000 observations on 1505 variables.
}
\usage{
data(dataPheWAS, package = 'EHR')
}
\description{
Simulated outcome data example from Phenome Wide Association Study (PheWAS) 
that examines associations between drug exposure and various phenotypes at follow-up 
after the drug exposure. The dataset includes 1505 variables: subject identification 
number ('id'), drug exposure ('exposure'), 3 demographic variables 
('age', 'race', 'gender'), and 1500 phenotypes.
}
\examples{
data(dataPheWAS)
}
\keyword{datasets}
