% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_data.R
\docType{data}
\name{dem}
\alias{dem}
\title{Simulated EHR data - Demographics.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 400 rows and 9 columns.}
\usage{
data(dem)
}
\description{
Simulated EHR data (no PHI constraints) featuring demographics data for ~500 patients.
}
\details{
Maintained by: Clara Marquardt
}
\keyword{datasets}
