% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_paste.R
\name{dt_paste}
\alias{dt_paste}
\title{Paste elements of a frequency table into a string.}
\usage{
dt_paste(list_dt, text, length = 10, text_col = NA, number_col = NA,
  line_sep = F)
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Paste elements of a frequency table into a string.
}
\examples{
TBC
}
