% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_replace.R
\name{dt_replace}
\alias{dt_replace}
\title{Perform a set of common in-place modifications on a data.table.}
\usage{
dt_replace(dt, mode, replace = "DEFAULT", col_name = names(dt))
}
\arguments{
\item{dt}{Data.table which is to be modified in place (data.table).}

\item{mode}{Modification which is to be performed ((i) set_na_zero (replace: Na, inf, -inf -> 0) (ii) set_zero_na (replace: 0 -> NA), (iii) set_missing_na (replace: " "* -> NA) (character).}

\item{replace}{Value used to replace the existing values (character) [Default: Mode-specific (see above)]}

\item{col_name}{Vector of column names which are to be modified (vector - character) [Default: All columns in the dt].}
}
\value{
data.table modified in place
}
\details{
Maintainer: Clara Marquardt
}
\examples{
sample_dt <- copy(test_dt)
sample_dt[, ':='(test_col=prediction, test_col_1=control_cat_3)]

## mode - set_na_zero
sample_dt[c(1,3,4), ':='(test_col=c(NA, -Inf), test_col_1=NA)]
print(sample_dt)
dt_replace(dt=sample_dt, mode="set_na_zero", col_name=c("test_col"))
print(sample_dt)
## mode - set_zero_na
sample_dt[c(1,3,4), ':='(test_col=0, test_col_1=0)]
print(sample_dt)
dt_replace(dt=sample_dt, mode="set_zero_na", col_name=c("test_col_1"))
print(sample_dt)

## mode - set_missing_na
sample_dt[, ':='(test_col=as.character(test_col), test_col_1=as.character(test_col_1))]
sample_dt[c(1,3,4), ':='(test_col=" ", test_col_1="     ")]
print(sample_dt)
dt_replace(dt=sample_dt, mode="set_missing_na")
print(sample_dt)
}
