% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_stat.R
\name{dt_stat}
\alias{dt_stat}
\title{Count number of observations in a data.table over a given timeframe/a given id list.}
\usage{
dt_stat(id_list, empi_list, date_list, dt, dt_date_var, timeframe_day = 365,
  buffer_day = 0, timeframe_hour = 24, buffer_hour = 0, exp_list = NA,
  exp_list_agg = NA, mode = "day", mode_direction = "backward")
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Count number of observations in a data.table over a given timeframe/a given id list.
}
\examples{
TBC
}
