% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EHR-package.R
\docType{package}
\name{EHR-package}
\alias{EHR}
\alias{EHR-package}
\title{Electronic Health Record (EHR) Data Processing and Analysis Tool}
\description{
The `EHR` package provides modules to perform diverse medication-related
studies using data from EHR databases.
}
\details{
Package functionality:
\itemize{
\item Process and analyze Electronic Health Record (EHR) data.

\item Implement modules to perform diverse medication-related studies using
data from EHR databases. Especially, the package includes modules to perform
pharmacokinetic/pharmacodynamic (PK/PD) analyses using EHRs, as outlined in
Choi et al. (2020).

\item Implement three statistical methods for Phenome Wide Association Study
(PheWAS). Contingency tables for many binary outcomes (e.g., phenotypes) and
a binary covariate (e.g., drug exposure) can be efficiently generated by
\code{\link{zeroOneTable}}, and three commonly used statistical methods to
analyze data for PheWAS are implement by \code{\link{analysisPheWAS}}.
}
}
\examples{
data(dataPheWAS)
demo.covariates <- c('id','exposure','age','race','gender')
phenotypeList <- setdiff(colnames(dd), demo.covariates)
tablePhenotype <- matrix(NA, ncol=4, nrow=length(phenotypeList), 
dimnames=list(phenotypeList, c("n.nocase.nonexp", "n.case.nonexp", 
"n.nocase.exp", "n.case.exp")))
for(i in seq_along(phenotypeList)) {
    tablePhenotype[i, ] <- zeroOneTable(dd[, 'exposure'], dd[, phenotypeList[i]])
}

## use small datasets to run this example
data(dataPheWASsmall)
## make dd.base with subset of covariates from baseline data (dd.baseline.small)
## or select covariates with upper code as shown below
upper.code.list <- unique(sub("[.][^.]*(.).*", "", colnames(dd.baseline.small)) )
upper.code.list <- intersect(upper.code.list, colnames(dd.baseline.small))
dd.base <- dd.baseline.small[, upper.code.list]
## perform regularized logistic regression to obtain propensity score (PS) 
## to adjust for potential confounders at baseline
phenos <- setdiff(colnames(dd.base), c('id', 'exposure'))
data.x <- as.matrix(dd.base[, phenos])
glmnet.fit <- glmnet::cv.glmnet(x=data.x, y=dd.base[,'exposure'],
                                family="binomial", standardize=TRUE,
                                alpha=0.1)
dd.base$PS <- c(predict(glmnet.fit, data.x, s='lambda.min'))
data.ps <- dd.base[,c('id', 'PS')]
dd.all.ps <- merge(data.ps, dd.small, by='id')  
demographics <- c('age', 'race', 'gender')
phenotypeList <- setdiff(colnames(dd.small), c('id','exposure','age','race','gender'))
## run with a subset of phenotypeList to get quicker results
phenotypeList.sub <- sample(phenotypeList, 10)
results.sub <- analysisPheWAS(method='firth', adjust='PS', Exposure='exposure',
                                                PS='PS', demographics=demographics, 
                                                phenotypes=phenotypeList.sub, data=dd.all.ps)
## run with the full list of phenotype outcomes (i.e., phenotypeList)
\donttest{
        results <- analysisPheWAS(method='firth', adjust='PS',Exposure='exposure',
                          PS='PS', demographics=demographics,
                          phenotypes=phenotypeList, data=dd.all.ps) 
}
}
\references{
\enumerate{
\item Development of a system for postmarketing population pharmacokinetic
and pharmacodynamic studies using real-world data from electronic health
records.\cr
Choi L, Beck C, McNeer E, Weeks HL, Williams ML, James NT, Niu X,
Abou-Khalil BW, Birdwell KA, Roden DM, Stein CM, Bejan CA, Denny JC,
Van Driest SL.\cr
Clin Pharmacol Ther. 2020 Apr;107(4):934-943. doi: 10.1002/cpt.1787.

\item Evaluating statistical approaches to leverage large clinical datasets
for uncovering therapeutic and adverse medication effects.\cr
Choi L, Carroll RJ, Beck C, Mosley JD, Roden DM, Denny JC, Van Driest SL.\cr
Bioinformatics. 2018 Sep 1;34(17):2988-2996. doi: 10.1093/bioinformatics/bty306.
}
}
\author{
\strong{Maintainer}: Leena Choi \email{naturechoi@gmail.com}

Authors:
\itemize{
  \item Cole Beck \email{cole.beck@vumc.org} (\href{https://orcid.org/0000-0002-6849-6255}{ORCID})
  \item Hannah Weeks \email{hannah.l.weeks@vanderbilt.edu}
  \item Elizabeth McNeer \email{elizabeth.mcneer@vumc.org}
}

}
\keyword{EHR}
\keyword{PheWAS}
\keyword{process}
