% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehR_cohort.R
\docType{data}
\name{ehR_cohort}
\alias{ehR_cohort}
\title{Simulated cohort & feature set.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 10000 rows and 14 columns.}
\usage{
data(ehR_cohort)
}
\description{
Cohort-style data.table to facilitate function/code testing and validation.
}
\details{
Maintained by: Clara Marquardt
}
\keyword{datasets}
