% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_add.R
\name{exist_add}
\alias{exist_add}
\title{Return object/object returned by expression if the object exists, return NULL otherwise.}
\usage{
exist_add(obj_exp)
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{

}
\examples{
TBC
}
