% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_table.R
\name{extract_table}
\alias{extract_table}
\title{Extract tables from a PDF journal article.}
\usage{
extract_table(temp_table)
}
\arguments{
\item{temp_table}{List with text rows extracted from PDF [list - character].}
}
\description{

}
