% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMedUtils.R
\name{extractMed-internal}
\alias{extractMed-internal}
\alias{getNote}
\alias{getDose}
\title{Internal functions for Extract-Med module}
\usage{
getNote(note)
}
\description{
These internal functions aid the main function used in the Extract-Med module (\code{extractMed}).
}
\details{
\code{getNote}:

\code{getDose}:
}
\keyword{internal}
