% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMed.R
\name{extractMed}
\alias{extractMed}
\title{Extract medication inforamtion from clinical notes}
\usage{
extractMed(note_fn, drugnames, drgunit, windowlength, max_edit_dist = 0, ...)
}
\arguments{
\item{note_fn}{File name(s) for the text files containing the clinical notes. Can be 
a character string for an individual note, or a vector or list of file names for 
multiple notes.}

\item{drugnames}{Vector of drug names for which dosing information should be extracted. 
Can include various forms (e.g., generic, brand name) as well as abbreviations.}

\item{drgunit}{Unit of the drug being extracted, e.g., 'mg'}

\item{windowlength}{Length of the search window (in characters) around the drug name in 
which to search for dosing entities}

\item{max_edit_dist}{Maximum edit distance allowed when attempting to extract \code{drugnames}. 
Allows for capturing misspelled drug name information.}

\item{...}{Additional arguments to \code{medExtractR}}
}
\value{
A data.frame with the extracted dosing information
}
\description{
This function is an interface to the \code{medExtractR} function within the \code{medExtractR} 
package, and allows drug dosing information to be extracted from free text sources, 
e.g., clinical notes.
}
