% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_dt.R
\name{feat_dt}
\alias{feat_dt}
\title{Generate, format and save tabulation of the most important predictors 
in a model.}
\usage{
feat_dt(feat_dt_raw, feat_imp_var_list, output_path, feat_lim = 200,
  add_dt = NULL)
}
\arguments{
\item{feat_dt_raw}{}

\item{feat_imp_var_list}{}

\item{output_path}{}

\item{feat_lim}{}
}
\description{

}
