% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_ext_stat.R
\name{feature_ext_stat}
\alias{feature_ext_stat}
\title{Generate descriptive stats for the features in a feature set.}
\usage{
feature_ext_stat(pred_set, outcome_var)
}
\arguments{
\item{pred_set}{}

\item{outcome_var}{}
}
\description{

}
