% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{form}
\alias{form}
\title{Convert a character vector into a formula object.}
\usage{
form(var_list, formula = TRUE)
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Convert a character vector into a formula object.
}
\examples{
TBC
}
