% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqUtils.R
\name{freq-internal}
\alias{freq-internal}
\alias{freqNum}
\alias{parseFreq}
\title{Internal functions to standardize frequency}
\description{
These internal functions aid the main functions when handling
frequency values.
}
\details{
\code{freqNum}: converts the frequency entity to numeric

\code{parseFreq}: standardizes the frequency entity
}
