% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_table_comb.R
\name{freq_table_comb}
\alias{freq_table_comb}
\title{Combine elements of a frequency table into a (character) string or formatted table.}
\usage{
freq_table_comb(freq_table, expl_text, element_number = 10, sort = TRUE,
  string_table = "string", item_sep = "/", text_col = NA,
  number_col = NA)
}
\arguments{
\item{freq_table}{frequency table (table(x) output).}

\item{expl_text}{text which is to be appended at the end of the string (e.g. "(% of diagnoses)") (character).}

\item{element_number}{number of rows of the frequency table which are to be included (integer) [default:10].}

\item{sort}{whether to sort the frequency table (descending order) (logical - TRUE/FALSE) [default: TRUE].}

\item{string_table}{whether to return a string or table (character - "string"/"table") [default: "string"].}

\item{item_sep}{symbol used to separate the elements in the string (character) [default: "//"].}

\item{text_col}{name of text column (character) [default: automatically identify column based on column types].}

\item{number_col}{name of number column (character) [default: automatically identify column based on column types].}
}
\value{
formatted (character) string or data.table.
}
\description{
Combine elements of a frequency table into a (character) string or formatted table.
}
\examples{
freq_table_raw  <- table(dia$dia_name)

freq_string <- freq_table_comb(freq_table=freq_table_raw, expl_text="\% of Principal Diagnoses", 
element_number=10, sort=TRUE, item_sep="/", string_table="string")
print(freq_string)

freq_table <- freq_table_comb(freq_table=freq_table_raw, expl_text="\% of Principal Diagnoses", 
element_number=10, sort=TRUE, item_sep="/", string_table="table")
print(freq_table)
}
