% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdzFreq.R
\name{freqNum}
\alias{freqNum}
\title{Convert Character Frequency to Numeric}
\usage{
freqNum(x)
}
\arguments{
\item{x}{character vector of extracted frequency values}
}
\value{
numeric vector
}
\description{
This function converts the frequency entity to numeric.
}
\examples{
f <- stdzFreq(c('in the morning', 'four times a day', 'with meals'))
freqNum(f)
}
