% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gagne_code_data.R
\docType{data}
\name{gagne_code}
\alias{gagne_code}
\title{Gagne Comorbididty Score Weights.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 33 rows and 7 columns.}
\source{
\url{http://scholar.harvard.edu/gagne/software/combined-comorbidity-score}
}
\usage{
data(gagne_code)
}
\description{
Weights used to derive Gagne Comborbidity Scores.
}
\keyword{datasets}
