% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gagne_score_calc.R
\name{gagne_score_calc}
\alias{gagne_score_calc}
\title{Calculate Gagne scores for a given set of patients over a given time period prior to t0.}
\usage{
gagne_score_calc(id_list, empi_list, date_list, dia_dt, timeframe_day = 365,
  buffer_day = 0, gagne_count_var = FALSE)
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Note: This function is designed to work with a Partners EHR master dataset, i.e. a data set
containing a dia_date_1 and gagne column.
}
\examples{
\dontrun{
gagne_score <- gagne_score_calc(
   dia$dia_id[1:1000],        
   dia$empi[1:1000],             
   dia$dia_date[1:1000],      
   dia,     
   timeframe_day=365, 
   buffer_day=1, 
   gagne_count_var=TRUE)
}
}
