% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_footnote_summary.R
\name{ggplot_footnote_summary}
\alias{ggplot_footnote_summary}
\title{A function that, given a ggplot object, list of summaries (each obtained 
using summary() in base R), a name for each summary object and an output 
file, returns (and saves) a version of the plot with a the summary 
(summaries) at the bottom of the plot.}
\usage{
ggplot_footnote_summary(ggplt, summary_list, summary_title_vec, ndigit = 3,
  font = "Helvetica", fontsize = 8, fontcolor = "#3A3F3F",
  output_file = NA)
}
\arguments{
\item{ggplt}{The ggplot to which the summary(summaries) are to be appended}

\item{summary_list}{The list of summaries obtained using summary()}

\item{summary_title_vec}{Vector of strings to be used as titles for the summary (summaries)}

\item{ndigit}{Decimal precision to use in writing the summary statistics}

\item{font}{fontfamily used in writing footnote, "Helvetica" by default}

\item{fontsize}{footnote text size, 8 by default}

\item{fontcolor}{color of footnote text, dark grey by default}

\item{output_file}{Path and filename to save resulting plot with footnote as (NA, means don't save - default)}
}
\description{
Author: Shreyas Lakhtakia, slakhtakia [at] bwh.harvard.edu
}
\examples{
gender_summary <- list(summary(dem[gender == "female", ]$age), summary(dem[gender == "male", ]$age))
age_distribution <- ggplot(data = dem, aes(x = age, color = factor(gender))) + geom_density()
ggplot_footnote_summary(ggplt = age_distribution, summary_list = gender_summary, summary_title_vec = c("Female", "   Male"), output_file = "age_distribution_by_gender.png")
}
