% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_line_plot.R
\name{hist_line_plot}
\alias{hist_line_plot}
\title{Generate and save a predicted vs. actual probability (i.e. calibration 
plot).}
\usage{
hist_line_plot(dt_list, outcome_var, cat_var, cat_num = 10,
  unit_desc = "deciles", plot_title = "", subtitle_list = list("", ""),
  return_data = FALSE, plot_point = TRUE, plot_id_line = TRUE,
  outcome_name = "outcome", output_path, output_name,
  hor_scaling_factor = 10)
}
\arguments{
\item{dt_list}{}

\item{outcome_var}{}

\item{cat_var}{}

\item{cat_num}{}

\item{unit_desc}{}

\item{plot_title}{}

\item{subtitle_list}{}

\item{return_data}{}

\item{plot_point}{}

\item{plot_id_line}{}

\item{outcome_name}{}

\item{output_path}{}

\item{output_name}{}

\item{hor_scaling_factor}{}

\item{"")}{}
}
\description{

}
