% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hochberg_adj_hyderabad.R
\name{hochberg_adj_hyderabad}
\alias{hochberg_adj_hyderabad}
\title{Perform a Hochberg correction of p-values to account for multiple hypothesis testing [Hochberg (1988) - Step Down (?) (Hyderabad Paper)].}
\usage{
hochberg_adj_hyderabad(p)
}
\arguments{
\item{p}{Vector of raw p-values [numeric].}
}
\description{
Perform a Hochberg correction of p-values to account for multiple hypothesis testing [Hochberg (1988) - Step Down (?) (Hyderabad Paper)].
}
