% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holm_adj.R
\name{holm_adj}
\alias{holm_adj}
\title{Perform a Holm correction of p-values to account for multiple hypothesis testing [Holm (1988) - Step Down].}
\usage{
holm_adj(p)
}
\arguments{
\item{p}{Vector of raw p-values [numeric].}
}
\description{
Perform a Holm correction of p-values to account for multiple hypothesis testing [Holm (1988) - Step Down].
}
