% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsId.R
\name{idCrosswalk}
\alias{idCrosswalk}
\title{Create ID Crosswalk}
\usage{
idCrosswalk(data, idcols)
}
\arguments{
\item{data}{list of data.frames}

\item{idcols}{list of character vectors, indicating ID columns
found in each data set given in \sQuote{data}}
}
\value{
crosswalk of ID columns and their de-identified versions
}
\description{
Link ID columns from multiple data sets. Currently restricted to
\sQuote{subject_id} and \sQuote{subject_mrn}. De-identified columns
are created to make a crosswalk.
}
\details{
\sQuote{subject_id} and \sQuote{subject_mrn} may occur multiple
times, but should have a one-to-one linkage defined by at least
one of the input data sets. A new visit number is generated for
each repeated \sQuote{subject_mrn}.
}
\examples{
\donttest{
data <-  list(demo_data, conc_data)
idcols <-  list(c('subject_id', 'subject_mrn'), 'subject_id')
idCrosswalk(data, idcols)
}
}
