% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_batch.R
\name{ingest_meta}
\alias{ingest_meta}
\title{Capture run-time stats for scripts executed in batch-mode.}
\usage{
ingest_meta(start_var_end, file_name_var, meta_batch_doc_arg = meta_batch_doc,
  data_sample_var = data_sample, date = as.character(format(Sys.time(),
  "\%d/\%m/\%Y_\%H:\%M")), source_date_time_var = source_date_time)
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Capture run-time stats for scripts executed in batch-mode.
}
\examples{
\dontrun{
 current_date_time <- as.character(format(Sys.time(), "\%d/\%m/\%Y_\%H:\%M"))
 data_sample       <- "test_cohort"
 doc_path          <- "test.csv"
 ingest_meta("start", file_name_var="test_script", meta_batch_doc_arg=doc_path, data_sample_var=data_sample,
   source_date_time_var=current_date_time) ## start of script
 ingest_meta("end", file_name_var="test_script", , meta_batch_doc_arg=doc_path, data_sample_var=data_sample,
   source_date_time_var=current_date_time) ## end of script
}
}
