% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_col.R
\name{insert_col}
\alias{insert_col}
\title{Insert a column into a data.table at a specified position.}
\usage{
insert_col(dt, col_vector, pos, name)
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Insert a column into a data.table at a specified position.
}
\examples{
TBC
}
