% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamMetadata.R
\docType{data}
\name{lam_metadata}
\alias{lam_metadata}
\title{Example of Metadata for Lamotrigine Data}
\format{
A data frame with 5 observations on the following variables.
\describe{
  \item{filename}{A character vector, filename for the clinical note}
  \item{pid}{A character vector, patient ID associated with the filename}
  \item{date}{A character vector, date associated with the filename}
  \item{note}{A character vector, note ID associated with the filename}
}
}
\usage{
lam_metadata
}
\description{
An example of the metadata needed for the \code{makeDose} and \code{collapseDose} functions.
}
\examples{
data(lam_metadata)
}
\keyword{datasets}
