% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastdoseUtils.R
\name{lastdose-internal}
\alias{lastdose-internal}
\alias{matchLastDose}
\alias{mergeLastDose}
\alias{checkTime}
\alias{checkDuration}
\title{Internal functions for lastdose}
\description{
These internal functions incorporate lastdose.
}
\details{
\code{matchLastDose}: match lastdose data to built data

\code{mergeLastDose}: merge lastdose data into built data

\code{checkTime}: check if lastdose expression is a time expression in the format HH:MM:SS

\code{checkDuration}: check if lastdose expression is a duration expression (number that is 
not in the HH:MM:SS format)
}
\keyword{internal}
