% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_function.R
\name{list_function}
\alias{list_function}
\title{List all the functions used in a R script.}
\usage{
list_function(filename, alphabetic = TRUE)
}
\arguments{
\item{filename}{Path to R script}
}
\value{
TBC
}
\description{
List all the functions used in a R script.
}
\examples{
TBC
}
