% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_space.R
\name{list_space}
\alias{list_space}
\title{Insert element - string - into a list.}
\usage{
list_space(list_name, argument = "space")
}
\arguments{
\item{list_name}{List name [character]}

\item{argument}{String to be inserted[character]}
}
\value{
List modified in place
}
\description{
Insert element - string - into a list.
}
\examples{
 sample_list <- list()
 sample_list$nfold <- 10
 list_space("sample_list")
}
