% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_function.R
\name{list_table}
\alias{list_table}
\title{A Cat Function}
\usage{
list_table(list_name, var_name, argument = "space", vital_signs_merge = F,
  vital_signs_old = NA)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
