% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dep_DEV.R
\name{load_dep}
\alias{load_dep}
\title{Load dependencies (DEV function - to be used only for internal testing and development).}
\usage{
load_dep()
}
\description{
Add any (temporary) dependencies as @import statements.
}
