% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_function.R
\name{make_function}
\alias{make_function}
\title{Generate a function programatically.}
\usage{
make_function(arg, body, env = parent.frame())
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Generate a function programatically.
}
\examples{
TBC
}
