% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDose.R
\name{makeDose}
\alias{makeDose}
\title{Make Dose Data}
\usage{
makeDose(x, noteMetaData, naFreq = "most")
}
\arguments{
\item{x}{data.frame}

\item{noteMetaData}{data.frame}

\item{naFreq}{Replacing missing frequencies with this value, or the most
common.}
}
\value{
A list containing two dataframes, one with the note level and one with the date level collapsed data.
}
\description{
Takes parsed and paired medication data, calculates dose intake and daily dose, and removes redundant information at the note and date level.
}
\details{
This function standardizes frequency, route, and duration entities. Dose amount, strength, and frequency entities are converted to numeric. Rows with only drug name and/or route are removed. If there are drug name changes in adjacent rows, these rows are collpased into one row if there are no conflicts. Missing strengths, dose amounts, frequencies, and routes are borrowed or imputed using various rules. Dose given intake and daily dose are calculated. Redundancies are removed at the date and note level.
}
