% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclapply_robust.R
\name{mclapply_robust}
\alias{mclapply_robust}
\title{mclapply function useable across Mac/Windows.}
\usage{
mclapply_robust(max_core = 8, quiet = TRUE, ...)
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
mclapply function useable across Mac/Windows.
}
\examples{
#' \dontrun{
	result_robust <- mclapply_robust(X=seq(1:100), FUN=function(x) x*3)
	result        <- mclapply(seq(1:100), function(x) x*3)
}
}
