% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_ci.R
\name{mean_ci}
\alias{mean_ci}
\title{Calculate the confidence interval for a sample mean or proportion.}
\usage{
mean_ci(var, ci, return_sd = TRUE, digit = 2, min_lim = NA,
  max_lim = NA, type = "mean")
}
\value{
Confidence interval (and optionally standard deviation) [character vector]
}
\description{
Calculate the confidence interval for a sample mean or proportion.
}
\examples{
TBC
}
