% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_ci_clust.R
\name{mean_ci_clust}
\alias{mean_ci_clust}
\title{Calculate the clustered confidence interval for a sample mean or proportion.}
\usage{
mean_ci_clust(var, cluster_var, ci, return_sd = TRUE, digit = 2,
  min_lim = NA, max_lim = NA, type = "perc_prop")
}
\value{
Confidence interval (and optionally standard deviation) [character vector]
}
\description{
Calculate the clustered confidence interval for a sample mean or proportion.
}
\examples{
TBC
}
