% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_ci.R
\name{median_ci}
\alias{median_ci}
\title{Calculate the confidence interval for a sample median.}
\usage{
median_ci(var, ci = 0.95, return_sd = TRUE, digit = 2)
}
\value{
Confidence interval [character vector]
}
\description{
Calculate the confidence interval for a sample median.
}
\examples{
TBC
}
