% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modUtils.R
\name{mod-internal}
\alias{mod-internal}
\alias{colflow_mod}
\alias{findMedFlowRow_mod}
\alias{flowData_mod}
\alias{concData_mod}
\alias{infusionData_mod}
\alias{resolveDoseDups_mod}
\title{Internal functions for EHR modules}
\description{
code{colflow_mod}: build a comparison of potentially duplicate flow data
}
\details{
code{findMedFlowRow_mod}: find row numbers for matching ID and time

code{flowData_mod}: correct flow data

code{concData_mod}: correct concentration data

code{infusionData_mod}: combine flow and MAR data; standardize hourly rate,
which may or may not include weight

code{resolveDoseDups_mod}: correct duplicated infusion and bolus doses
}
\keyword{internal}
