% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_auc.R
\name{model_auc}
\alias{model_auc}
\title{Simple auc calculation.}
\usage{
model_auc(pred, outcome, auc_digits = 6)
}
\arguments{
\item{pred}{}

\item{outcome}{}

\item{auc_digits}{}
}
\description{
Generate auc for e.g. test and train predictions/outcomes 
at the same time.
}
