% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_merge.R
\name{multi_merge}
\alias{multi_merge}
\title{Merge together multiple data.tables.}
\usage{
multi_merge(data_list, key_var, allow.cartesian = FALSE)
}
\arguments{
\item{data_list}{List of data.tables which are to be merged (list - data.table).}

\item{key_var}{Vector of variable names on the basis of which the data.tables are to be merged (vector - character).}

\item{allow.cartesian}{See ?data.table (logical - TRUE/FALSE) [default: FALSE].}
}
\value{
Merged data.table.
}
\description{
Sequentially merge a list of data.tables on a given set of variables (inner-join).
}
\details{
Maintained by: Clara Marquardt
}
\examples{
data_list   <- list(ed[1:100], dia[1:100], dem[1:100])
dt_merged <- multi_merge(data_list=data_list, key_var=c("empi"))
print(str(dt_merged))
}
