% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiformat_glm.R
\name{multiformat_glm}
\alias{multiformat_glm}
\title{Given a list of logistic regressions, *performed using glm()* format, 
return and optionally, write the results side-by-side to csv.}
\usage{
multiformat_glm(glm_list, output_file = NA, formula_list = NA,
  title_list = "", ndigit = 5)
}
\arguments{
\item{glm_list}{a list() of glm model objects}

\item{output_file}{the file path (with ".csv" extension) to write the results to, if no value provided, no output file will be created}

\item{formula_list}{optional list of custom formulae to provide to print in results (code pulls regression formula call by default)}

\item{title_list}{optional list of titles for each regression in the list to write atop output file (blank by default)}

\item{ndigit}{level of precision in output, 5 by default}
}
\value{
matrix containing formatted results from the provided regression list
}
\description{
Inspired by: 
https://www.r-bloggers.com/export-r-results-tables-to-excel-please-dont-kick-me-out-of-your-club/
Author: Shreyas Lakhtakia (slakhtakia [at] bwh.harvard.edu)
}
\examples{
dem[, died := 0, ]
dem[tolower(vital_status) \%like\% "deceased", died := 1, ]
glm_age             <- glm(died ~ age, data = dem)
glm_age_gender      <- glm(died ~ age + gender, data = dem)
glm_age_gender_race <- glm(died ~ age + gender + race, data = dem)
glm_list            <- list(glm_age, glm_age_gender, glm_age_gender_race)
multiformat_glm(glm_list = glm_list)
}
