% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiformat_plm.R
\name{multiformat_plm}
\alias{multiformat_plm}
\title{Given a list of OLS regressions on panel data, *performed using plm()* 
(in order to obtain clustered SEs), format, return and optionally, write 
the results side-by-side in a csv.}
\usage{
multiformat_plm(plm_list, output_file = NA, formula_list = NA,
  title_list = "", ndigit = 5)
}
\arguments{
\item{plm_list}{a list() of plm model objects}

\item{output_file}{the file path (with ".csv" extension) to write the results to, if no value provided, no output file will be created}

\item{formula_list}{optional list of custom formulae to provide to print in results (code pulls regression formula call by default)}

\item{title_list}{optional list of titles for each regression in the list to write atop output file (blank by default)}

\item{ndigit}{level of precision in output, 5 by default}
}
\value{
matrix containing formatted results from the provided regression list
}
\description{
Inspired by: 
https://www.r-bloggers.com/export-r-results-tables-to-excel-please-dont-kick-me-out-of-your-club/
Author: Shreyas Lakhtakia (slakhtakia [at] bwh.harvard.edu)
}
\examples{
TBC
}
