% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymerge.R
\name{mymerge}
\alias{mymerge}
\title{Merge list of data.tables on a given set of variables.}
\usage{
mymerge(x, y, var_list = cohort_key_var)
}
\arguments{
\item{x, y}{Names of DT [object]}

\item{by}{Names of variables on which to merge  [character vector]}
}
\value{
Merged data.table
}
\description{
Merge list of data.tables on a given set of variables.
}
\examples{
TBC
}
