% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_outcome.R
\name{norm_outcome}
\alias{norm_outcome}
\title{Generate a Kling Index from a list of outcomes.}
\usage{
norm_outcome(df, outcome_list, include_missing = TRUE,
  assignment_var = "treatment")
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Generate a Kling Index from a list of outcomes.
}
\examples{
TBC
}
