% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.R
\name{pad}
\alias{pad}
\title{Pad elements of list with 0s to bring all elements to a fixed length.}
\usage{
pad(topad, num, method = "front", replace = "0")
}
\arguments{
\item{topad}{List of elements to pad [list]}

\item{num}{Final length [integer]}

\item{method}{Whether to pad front or back [character]}

\item{replace}{What to pad with [character]}
}
\value{
TBC
}
\description{
Pad elements of list with 0s to bring all elements to a fixed length.
}
\examples{
TBC
}
