% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_arg.R
\name{parse_arg}
\alias{parse_arg}
\title{Parse and assign (in parent namespace) the arguments in a function call.}
\usage{
parse_arg(x)
}
\arguments{
\item{x}{Function call (see examples) [character].}
}
\description{
Parse and assign (in parent namespace) the arguments in a function call.
}
\examples{
\dontrun{
 parse_arg(quote(test_function(arg1="xxxx", arg2=object_name, arg3="xxxx")))
}
}
