% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_date.R
\name{parse_date}
\alias{parse_date}
\title{Automatically recognise and parse date variables.}
\usage{
parse_date(dt_vector, column_name_list,
  column_name_list_new = column_name_list)
}
\arguments{
\item{dt_vector}{}
}
\description{
/
}
