% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMedex.R
\name{parseMedEx}
\alias{parseMedEx}
\title{Parse MedEx NLP Output}
\usage{
parseMedEx(fh)
}
\arguments{
\item{fh}{File name for a single file containg MedEx output.}
}
\value{
A data.table object with columns for filename, drugname, strength, dose, route, 
freq, and duration. The filename contains the file name corresponding to the clinical 
note. Each of the entity columns are of the format 
"extracted expression::start position::stop position".
}
\description{
Take files with the raw medication extraction output generated by the MedEx 
natural language processing system and converts it into a standardized format.
}
\details{
Output from different medication extraction systems is formatted in different ways.
In order to be able to process the extracted information, we first need to convert 
the output from different systems into a standardized format. Extracted expressions 
for various drug entities (e.g., drug name, strength, frequency, etc.) each receive 
their own column formatted as "extracted expression::start position::stop position". 
MedEx output files anchor extractions to a specific drug name extraction. In the case 
where multiple entities are extracted for the same drug name mention, the entities are 
separated into adjacent rows and the drug name is duplicated.
}
