% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perc.R
\name{perc}
\alias{perc}
\title{Calculate a percentage.}
\usage{
perc(num, denom, digit = 1)
}
\arguments{
\item{num}{Numerator [numeric]}

\item{denom}{Denominator [numeric]}

\item{digit}{Number of digits to round to [integer]}
}
\value{
Percentage [numeric]
}
\description{
Calculate a percentage.
}
\examples{
TBC
}
