% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_param_test.R
\name{perm_param_test}
\alias{perm_param_test}
\title{Permutation based F/t-test}
\usage{
perm_param_test(df, assignment_var, group_var,
  permutation = ExpLearning.Default$permutation,
  cluster_var = ExpLearning.Default$cluster_var,
  max_core = ExpLearning.Default$max_core, test = ExpLearning.Default$test)
}
\value{
TBC
}
\description{
Permutation based F/t-test
}
\examples{
TBC
}
