% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_by_factor.R
\name{plot_calibration_by_risk_quantile_by_factor}
\alias{plot_calibration_by_risk_quantile_by_factor}
\title{Calibration by factor}
\usage{
plot_calibration_by_risk_quantile_by_factor(data, output_file, outcome_col_name,
  quantile_col_name, cluster_by_col_name, plot_by_col_name, xlabel, ylabel,
  legend_label, color_palette = "Oranges", SE_line = TRUE,
  SE_style = "ribbon", ymin = 0, make_footnote = TRUE)
}
\arguments{
\item{data}{data.table object containing data to be plotted. (data table)}

\item{output_file}{filepath to which we write plot. (character)}

\item{outcome_col_name}{column name of outcome for which we want the mean (by factor and risk quantile) plotted on y-axis. (character)}

\item{quantile_col_name}{column name of risk quantile to go on x-axis (could be something like y_hat_percentile). (character)}

\item{cluster_by_col_name}{column name of grouping for which we want to cluster standard errors (typically something like empi). (character)}

\item{plot_by_col_name}{column name of factor variable for which we want separate lines. (character)}

\item{xlabel}{x-axis label. (character)}

\item{ylabel}{y-axis label. (character)}

\item{legend_label}{legend title, should describe element passed to 'plot_by_col_name'. (character)}

\item{color_palette}{a single string referring to a palette (see http://colorbrewer2.org/#type=sequential&scheme=BuGn&n=3 for details), or a 
vector of hex color codes matching the number of levels in plot_by_col_name. (character)}

\item{SE_line}{whether or not to include standard error band. (boolean)}

\item{SE_style}{style of standard error band (either 'ribbon' or 'line'). (character)}

\item{ymin}{minimum of y-axis. (numeric)}

\item{make_footnote}{whether or not to include a footnote on the plot (boolean)}
}
\description{
Plot calibration curve by factor column with risk quantile as x-axis
}
\examples{
\dontrun{
data <- copy(ehR_cohort)
data[, prediction_decile := add_quantile(prediction, 10)]
plot_calibration_by_risk_quantile_by_factor(data = data, output_file = 'calibration_by_factor.png', outcome_col_name = 'treatment',
                                            quantile_col_name = 'prediction_decile', cluster_by_col_name = 'empi', plot_by_col_name = 'test',
                                            xlabel = 'Predicted risk (decile)', ylabel = 'Proportion treated', legend_label = 'Tested')
}
}
