% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_marginal.R
\name{plot_marginal}
\alias{plot_marginal}
\title{Annotate a plot with one/two side plots.}
\usage{
plot_marginal(main_plot, hor_plot, vert_plot = NA, layout_width = c(4, 0.5),
  layout_height = c(0.5, 0.5, 2.5), plot_count = 2, title_obj = "",
  subtitle_obj = "")
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Note: This function is designed to work with a Partners EHR master dataset, i.e. a data set
containing a dia_date_1 and gagne column.
}
\examples{
TBC
}
