% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_memory.R
\name{print_memory}
\alias{print_memory}
\title{Print table with memory usage of objects in namespace.}
\usage{
print_memory(namespace = ls(parent.env(environment())))
}
\arguments{
\item{namespace}{Namespace}
}
\value{
TBC
}
\description{
Print table with memory usage of objects in namespace.
}
\examples{
print_memory()
}
