% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_r2.R
\name{pseudo_r2}
\alias{pseudo_r2}
\title{Calculate the pseudo-R2 value for a logistic model.}
\usage{
pseudo_r2(model)
}
\arguments{
\item{model}{Logistic regression model [model].}
}
\description{
Calculate the pseudo-R2 value for a logistic model.
}
