% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idUtils.R
\name{pullRealId}
\alias{pullRealId}
\title{Pull Real ID}
\usage{
pullRealId(dat, xwalk = NULL)
}
\arguments{
\item{dat}{a data.frame}

\item{xwalk}{a data.frame, providing linkage for each ID; if NULL, the
crosswalk will be pulled from the \sQuote{pkxwalk} option, or otherwise fail}
}
\value{
The modified data.frame
}
\description{
Replace de-identified IDs with identified version pulled from a crosswalk.
}
