% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRDS_merge.R
\name{readRDS_merge}
\alias{readRDS_merge}
\title{Read in (a) a single or (b) a list of RDS files (which are to be 
merged).}
\usage{
readRDS_merge(file_name_list, nested = FALSE)
}
\arguments{
\item{file_name_list}{}
}
\description{
/
}
