% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransformation.R
\name{readTransform}
\alias{readTransform}
\title{Read and Transform}
\usage{
readTransform(file, ...)
}
\arguments{
\item{file}{filename of a CSV file}

\item{\dots}{additional information passed to \code{\link{dataTransformation}}}
}
\value{
The modified data.frame
}
\description{
Convenience function for reading in a CSV file, and making small
modifications to a data.frame.
}
\details{
If \code{\link{read.csv}} needs additional arguments (or the file is in a
different format), the user should load the data first, then directly call
\code{\link{dataTransformation}}.
}
