% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na_zero_missing.R
\name{replace_na_zero_missing}
\alias{replace_na_zero_missing}
\title{Perform a set of common in-place modifications on a data.table.}
\usage{
replace_na_zero_missing(data, replace, replace_with = "DEFAULT",
  col = names(data))
}
\arguments{
\item{data}{Data.table which is to be modified in place (data.table).}

\item{replace}{Value used to replace the existing values (character) [Default: Mode-specific (see above)]}

\item{col}{Vector of column names which are to be modified (vector - character) [Default: All columns in the data].}

\item{mode}{Modification which is to be performed ((i) set_na_zero (replace: Na, inf, -inf -> 0) (ii) set_zero_na (replace: 0 -> NA), (iii) set_missing_na (replace: " "* -> NA) (character).}
}
\value{
data.table modified in place
}
\details{
Maintained by: Clara Marquardata
}
\examples{
sample_data <- copy(test_data)
sample_data[, ':='(test_col=prediction, test_col_1=control_cat_3)]

## mode - set_na_zero
sample_data[c(1,3,4), ':='(test_col=c(NA, -Inf), test_col_1=NA)]
print(sample_data)
replace_na_zero_missing(data=sample_data, replace="set_na_zero", col=c("test_col"))
print(sample_data)
## mode - set_zero_na
sample_data[c(1,3,4), ':='(test_col=0, test_col_1=0)]
print(sample_data)
replace_na_zero_missing(data=sample_data, replace="set_zero_na", col=c("test_col_1"))
print(sample_data)

## mode - set_missing_na
sample_data[, ':='(test_col=as.character(test_col), test_col_1=as.character(test_col_1))]
sample_data[c(1,3,4), ':='(test_col=" ", test_col_1="     ")]
print(sample_data)
replace_na_zero_missing(data=sample_data, replace="set_missing_na")
print(sample_data)
}
