% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_mult.R
\docType{data}
\name{return_mult}
\alias{return_mult}
\title{Store multiple arguments returned from function as separate objects.}
\format{An object of class \code{result} of length 1.}
\usage{
return_mult
}
\arguments{
\item{topad}{List of elements to pad [list]}

\item{num}{Final length [integer]}

\item{method}{Whether to pad front or back [character]}

\item{replace}{What to pad with [character]}
}
\value{
TBC
}
\description{
Store multiple arguments returned from function as separate objects.
}
\examples{
\dontrun{
within function_test: return(list(obj_1,obj_2, obj_3))
outside of function: return_mult[obj_1, obj_2, obj_3] <- function_test()
}
}
\keyword{datasets}
