% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocr_basic.R
\name{rocr_basic}
\alias{rocr_basic}
\title{Generate and save a (test and train) ROC curve.}
\usage{
rocr_basic(pred_list, outcome_list, model_name, title, path)
}
\arguments{
\item{pred_list}{}

\item{outcome_list}{}

\item{model_name}{}

\item{title}{}

\item{path}{}
}
\description{

}
