% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modGenPK.R
\name{run_Build_PK}
\alias{run_Build_PK}
\title{Build-PK Module}
\usage{
run_Build_PK(
  dose,
  dose.columns = list(),
  conc = NULL,
  conc.columns = list(),
  doseFreq = NULL,
  dosePriorWindow = 7,
  postWindow = NA,
  date.format = "\%m/\%d/\%y \%H:\%M:\%S",
  tz = ""
)
}
\arguments{
\item{dose}{dose data, a filename (CSV, RData, RDS), or a correctly formatted data.frame}

\item{dose.columns}{a named list that should specify columns in dose data;
\sQuote{id} is required. Dose data can be provided with the \sQuote{datetime},
\sQuote{dose}, and \sQuote{rate} variables. The date-time variable can be specified as a
single date-time variable (datetime = \sQuote{date_time}) or two variables
holding date and time separately (e.g., datetime = c(\sQuote{Date}, \sQuote{Time})).
\sQuote{druglevel} should be present if a separate concentration data set is not provided.}

\item{conc}{concentration data, a filename (CSV, RData, RDS), or a
correctly formatted data.frame. If NULL, concentration data should be present in the dose data.}

\item{conc.columns}{a named list that should specify columns in concentration
data; \sQuote{id}, \sQuote{datetime}, \sQuote{druglevel} are required.}

\item{doseFreq}{Provides dose schedule interval for repeated dosing case. A value of 2 would
indicate two doses per day. The default (NULL) will not use repeated dosing.
Alternatively \sQuote{II} can be specified as a dose.column. II=12 would
indicate two doses per day.}

\item{dosePriorWindow}{Dose data is merged with drug level data. This value sets the
time frame window with the number of days prior to the first drug level data; defaults to 7.}

\item{postWindow}{Data is merged with drug level data. This postWindow can set the end time
for the drug level data, being the number of days after the first drug level data. The
default (NA) will use the date of the last drug level data.}

\item{date.format}{output format for \sQuote{date} variable}

\item{tz}{time zone for date-time variables}
}
\value{
PK data set
}
\description{
This module builds PK data for properly formatted dose and concentration data.
}
