% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modPK.R
\name{run_Build_PK_IV}
\alias{run_Build_PK_IV}
\title{Build PK Data}
\usage{
run_Build_PK_IV(
  conc,
  dose,
  lab.dat = NULL,
  lab.vars = NULL,
  demo.list = NULL,
  demo.vars = NULL,
  demo.abbr = NULL,
  pk.vars,
  drugname,
  check.path,
  missdemo_fn = "-missing-demo",
  faildupbol_fn = "DuplicateBolus-",
  date.format = "\%m/\%d/\%y \%H:\%M:\%S",
  date.tz = "America/Chicago"
)
}
\arguments{
\item{conc}{concentration data, the output of \code{\link{run_DrugLevel}}}

\item{dose}{dose data, the output of \code{\link{run_MedStrI}}}

\item{lab.dat}{lab data, if available}

\item{lab.vars}{variables to include from lab data}

\item{demo.list}{demographic information, if available}

\item{demo.vars}{variables to include from demographic data}

\item{demo.abbr}{character vector used to rename/abbreviate demo variables}

\item{pk.vars}{variables to include from PK data}

\item{drugname}{drug of interest, included in filename of check files}

\item{check.path}{path to \sQuote{check} directory, where check files are
created}

\item{missdemo_fn}{filename for checking NA frequency among demographic data}

\item{faildupbol_fn}{filename for duplicate bolus data}

\item{date.format}{output format for \sQuote{date} variable}

\item{date.tz}{output time zone for \sQuote{date} variable}
}
\value{
PK data set
}
\description{
This module will build PK IV data.
}
