% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkDosingInterval.R
\name{run_Build_PK_Oral}
\alias{run_Build_PK_Oral}
\title{PK oral data}
\usage{
run_Build_PK_Oral(
  x,
  idCol = "id",
  dtCol = "dt",
  doseCol = "dose",
  concCol = "conc",
  ldCol = NULL,
  first_interval_hours = 336,
  imputeClosest = NULL
)
}
\arguments{
\item{x}{data.frame}

\item{idCol}{data.frame id column name}

\item{dtCol}{data.frame date column name}

\item{doseCol}{dose column name}

\item{concCol}{concentration column name}

\item{ldCol}{last-dose time column name}

\item{first_interval_hours}{number of hours before first imputed time point}

\item{imputeClosest}{columns to impute with closest value}
}
\value{
data.frame
}
\description{
This module builds oral data for PK analysis.
}
