% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modDemo.R
\name{run_Demo}
\alias{run_Demo}
\title{Run Demographic Data}
\usage{
run_Demo(demo.path, toexclude, demo.mod.list)
}
\arguments{
\item{demo.path}{filename of a lab file (stored as RDS)}

\item{toexclude}{expression that should evaluate to a logical, indicating if
the observation should be excluded}

\item{demo.mod.list}{list of expressions, giving modifications to make}
}
\value{
list with two components
  \item{demo}{demographic data}
  \item{exclude}{vector of excluded visit IDs}
}
\description{
This module will load and modify demographic data.
}
