% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modDrugLevel.R
\name{run_DrugLevel}
\alias{run_DrugLevel}
\title{Run Drug Level Data}
\usage{
run_DrugLevel(
  conc.path,
  conc.select,
  conc.rename,
  conc.mod.list = list(mod_id_event = expression(paste(mod_id_visit, event, sep =
    "_"))),
  samp.path = NULL,
  samp.mod.list = list(mod_id_event = expression(paste(mod_id_visit, samp, sep = "_"))),
  check.path,
  failmiss_fn = "MissingConcDate-",
  multsets_fn = "multipleSetsConc-",
  faildup_fn = "DuplicateConc-",
  drugname,
  LLOQ,
  demo.list = NULL
)
}
\arguments{
\item{conc.path}{filename of concentration data (stored as RDS)}

\item{conc.select}{columns to select from concentration data}

\item{conc.rename}{new column names for concentration data}

\item{conc.mod.list}{list of expressions, giving modifications to make}

\item{samp.path}{filename of data with sampling time (stored as RDS)}

\item{samp.mod.list}{list of expressions, giving modifications to make}

\item{check.path}{path to \sQuote{check} directory, where check files are
created}

\item{failmiss_fn}{filename for data missing concentration date}

\item{multsets_fn}{filename for data with multiple concentration sets}

\item{faildup_fn}{filename for data with duplicate concentration observations}

\item{drugname}{drug of interest, included in filename of check files}

\item{LLOQ}{lower limit of concentration values; values below this are invalid}

\item{demo.list}{demographic information; if available, concentration records
must have a valid demo record}
}
\value{
drug-level data set
}
\description{
This module will load and modify drug-level data.
}
