% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modLabs.R
\name{run_Labs}
\alias{run_Labs}
\title{Run Lab Data}
\usage{
run_Labs(lab.path, lab.select, lab.mod.list)
}
\arguments{
\item{lab.path}{filename of a lab file (stored as RDS)}

\item{lab.select}{columns to select}

\item{lab.mod.list}{list of expressions, giving modifications to make}
}
\value{
lab data set
}
\description{
This module will load and modify laboratory data.
}
