% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modMedStr3.R
\name{run_MedStrIII}
\alias{run_MedStrIII}
\title{Run Str Data III}
\usage{
run_MedStrIII(
  dose.path,
  dose.columns = list(),
  req.vals = list(),
  check.path = NULL,
  drugname = NULL,
  conflict.window = 60
)
}
\arguments{
\item{dose.path}{filename of dose data (CSV, RData, RDS), or data.frame}

\item{dose.columns}{a named list that should specify columns in dose data; \sQuote{id},
\sQuote{datetime} and \sQuote{dose} are required.
\sQuote{datetime} is date and time for data measurement, which can refer to a single
date-time variable (datetime = \sQuote{date_time}) or two variables holding date and
time separately (e.g., datetime = c(\sQuote{Date}, \sQuote{Time})). \sQuote{duration} is optional
and should be given in seconds (e.g., 3600 would equate to 1 hour) when present.}

\item{req.vals}{a named list that can be used to filter data. As an example,
req.vals(unit = \sQuote{mg}, route = \sQuote{iv}) would indicate the dose data has a \sQuote{unit}
and \sQuote{route} column. In this case any observation with unit != mg or any route != iv
would be excluded.}

\item{check.path}{path to \sQuote{check} directory, where check files are
created. The default (NULL) will not produce any check files.}

\item{drugname}{drug of interest, included in filename of check files. The default (NULL)
will produce filenames without drugname included.}

\item{conflict.window}{Time window (in minutes) to consider nearby doses. The default is 60 minutes.
Equal doses within the window will be considered duplicats and removed. Unequal doses
will be also be excluded and saved to a check file (if \sQuote{check.path} is set).}
}
\value{
structured data set
}
\description{
This module will load and modify structured intravenous (IV) medication data. It expects actual
dose amount given. If duration is provided, rate will be calculated.
}
\details{
See EHR Vignette for Structured Data.
}
\examples{
# dose data for single user
doseData <- structure(list(id = c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1),
  dt = c("2020-01-01 08:00", "2020-01-02 08:00", "2020-01-02 20:00",
    "2020-01-02 20:00", "2020-01-03 08:00", "2020-01-05 08:00", "2020-01-05 08:00",
    "2020-01-05 08:00", "2020-01-05 20:00", "2020-01-06 08:00", "2020-01-06 08:00",
    "2020-01-06 20:00", "2020-01-07 20:00", "2020-01-07 20:00"),
  dose = c(30, 10, 20, 30, 10, 10, 20, 30, 10, 20, 10, 10, 20, 20),
  dur = c(3600, 7200, 1800, 3600, 3600, 7200, 1800, 3600, 7200, 1800, 1800, 1800, 1800, 3600)),
  class = "data.frame", row.names = c(NA, -14L))

run_MedStrIII(dose.path = doseData,
  dose.columns = list(id = 'id', datetime = 'dt', dose = 'dose', duration = 'dur'))

}
