% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_missing_na.R
\name{set_missing_na}
\alias{set_missing_na}
\title{Replace (in place) empty values ("[ ]*" or "") in data.table with another value.}
\usage{
set_missing_na(dt, replace = NA, subset_col = names(dt))
}
\arguments{
\item{replace}{Value with which to replace empty values  [any]}

\item{name}{Name of data.table [character]}
}
\value{
Data.table modified in place
}
\description{
Replace (in place) empty values ("[ ]*" or "") in data.table with another value.
}
\examples{
TBC
}
