% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_na_zero.R
\name{set_na_zero}
\alias{set_na_zero}
\title{Replace (in place) NAs/+inf/-inf in data.table with another value.}
\usage{
set_na_zero(dt, replace = 0, subset_col = names(dt))
}
\arguments{
\item{replace}{Value with which to replace Nas/+inf/-inf  [any]}

\item{name}{Name of data.table [character]}
}
\value{
Data.table modified in place
}
\description{
Replace (in place) NAs/+inf/-inf in data.table with another value.
}
\examples{
TBC
}
