% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_zero_na.R
\name{set_zero_na}
\alias{set_zero_na}
\title{Replace (in place) zeros in in data.table with another value.}
\usage{
set_zero_na(dt, replace = NA)
}
\arguments{
\item{replace}{Value with which to replace 0s  [any]}

\item{name}{Name of data.table [character]}
}
\value{
Data.table modified in place
}
\description{
Replace (in place) zeros in in data.table with another value.
}
\examples{
TBC
}
