% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorten_label.R
\name{shorten_label}
\alias{shorten_label}
\title{Shorten a character column of a data.table/a character vector.}
\usage{
shorten_label(DT, colname_list = NA, max_length = 20, beg_fraction = 0.8,
  in_place = TRUE, new_colname_list = NA)
}
\arguments{
\item{DT}{data.table or character vector [data.table / character vecor]}

\item{colname_list}{List of column names which are to be processed (NA if processing a vector)[character list]}

\item{max_length}{Max character length of shortened labels [integer]}

\item{in_place}{Modify in place,i.e. replace original labels [logical]}

\item{new_colname_list}{List of new column names to be generated if in_place==False (name of new vector object 
to be created if parsing a vector) [character list / vector]}

\item{beg_fraction:}{Number of characters to to be kept from the label's beginning (as a % of 
max_length characters)  [numeric]}
}
\value{
TBC
}
\description{
Shorten a character column of a data.table/a character vector.
}
\examples{
temp <- data.table::copy(dia)
shorten_label(temp, list("dia_name"), new_colname_list=list("dia_name_short"), in_place=FALSE)
}
