% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_star.R
\name{sign_star}
\alias{sign_star}
\title{Determine the significance (star format) of a p-value.}
\usage{
sign_star(p_value)
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Determine the significance (star format) of a p-value.
}
\examples{
TBC
}
