% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdzDose.R
\name{stdzDose}
\alias{stdzDose}
\title{Standardize Dose Entity}
\usage{
stdzDose(x)
}
\arguments{
\item{x}{character vector of extracted dose values}
}
\value{
numeric vector
}
\description{
This function standardizes the dose entity.
}
\details{
Some dose strings may include multiple values and additional interpretation
may be needed. For example \sQuote{2-1} likely indicates a dose of 2 followed
by a dose of 1. Currently it would be converted to the average of 1.5.
}
\examples{
stdzDose(c('one tablet', '1/2 pill', '1-3 tabs'))
}
