% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdzDoseChange.R
\name{stdzDoseChange}
\alias{stdzDoseChange}
\title{Standardize Dose Change Entity}
\usage{
stdzDoseChange(x)
}
\arguments{
\item{x}{character vector of extracted dose change values}
}
\value{
character vector
}
\description{
This function standardizes the dose change entity.
}
\examples{
stdzDoseChange(c('decreasing','dropped','increased'))
}
