% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdzDoseSchedule.R
\name{stdzDoseSchedule}
\alias{stdzDoseSchedule}
\title{Standardize Dose Schedule Entity}
\usage{
stdzDoseSchedule(x)
}
\arguments{
\item{x}{character vector of extracted dose schedule values}
}
\value{
character vector
}
\description{
This function standardizes the dose schedule entity.
}
\examples{
stdzDoseSchedule(c('tapered','weaned','TAPER'))
}
