% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdzDuration.R
\name{stdzDuration}
\alias{stdzDuration}
\title{Standardize Duration Entity}
\usage{
stdzDuration(x)
}
\arguments{
\item{x}{character vector of extracted duration values}
}
\value{
character vector
}
\description{
This function standardizes the duration entity.
}
\examples{
stdzDuration(c('1 month', 'three days', 'two-weeks'))
}
