% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdzFreq.R
\name{stdzFreq}
\alias{stdzFreq}
\title{Standardize Frequency Entity}
\usage{
stdzFreq(x)
}
\arguments{
\item{x}{character vector of extracted frequency values}
}
\value{
character vector
}
\description{
This function standardizes the frequency entity.
}
\examples{
stdzFreq(c('in the morning', 'four times a day', 'with meals'))
}
