% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdzRoute.R
\name{stdzRoute}
\alias{stdzRoute}
\title{Standardize Route Entity}
\usage{
stdzRoute(x)
}
\arguments{
\item{x}{character vector of extracted route values}
}
\value{
character vector
}
\description{
This function standardizes the route entity.
}
\examples{
stdzRoute(c('oral', 'po', 'subcut'))
}
