% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringify_frequency.R
\name{stringify_frequency}
\alias{stringify_frequency}
\title{Combine elements of a frequency table into a (character) string or formatted table.}
\usage{
stringify_frequency(data, expl_text, element_number = 10, sort = TRUE,
  string_table = "string", item_sep = "/", text_col = NA,
  number_col = NA, ndigit = 3)
}
\arguments{
\item{data}{frequency table (table(x) output).}

\item{expl_text}{text which is to be appended at the end of the string (e.g. "(Percent of diagnoses)") (character).}

\item{element_number}{number of rows of the frequency table which are to be included (integer) [default:10].}

\item{sort}{whether to sort the frequency table (descending order) (logical - TRUE/FALSE) [default: TRUE].}

\item{string_table}{whether to return a string or table (character - "string"/"table") [default: "string"].}

\item{item_sep}{symbol used to separate the elements in the string (character) [default: "//"].}

\item{text_col}{name of text column (character) [default: automatically identify column based on column types].}

\item{number_col}{name of number column (character) [default: automatically identify column based on column types].}
}
\value{
formatted (character) string or data.table.
}
\description{
Combine elements of a frequency table into a (character) string or formatted table.
}
\details{
Maintained by: Clara Marquardt
}
\examples{
data_raw  <- prop.table(table(dia$dia_name))
freq_string <- stringify_frequency(data=data_raw, expl_text="Percent of Principal Diagnoses", 
element_number=10, sort=TRUE, item_sep="/", string_table="string", ndigit=3)
print(freq_string)
data <- stringify_frequency(data=data_raw, expl_text="Percent of Principal Diagnoses", 
element_number=10, sort=TRUE, item_sep="/", string_table="table", ndigit=3)
print(data)
}
