% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_mult_function.R
\name{[<-.result}
\alias{[<-.result}
\title{Store multiple arguments returned from function as separate objects - function.}
\usage{
\method{[}{result}(x, ...) <- value
}
\value{
TBC
}
\description{
Store multiple arguments returned from function as separate objects - function.
}
\examples{
\dontrun{
within function_test: return(list(obj_1,obj_2, obj_3))
outside of function: return_mult[obj_1, obj_2, obj_3] <- function_test()
}
}
