% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemfit_unrelated.R
\name{systemfit_unrelated}
\alias{systemfit_unrelated}
\title{Generate a systemfit object from a truly unrelated system of regressions (equivalent to systemfit vs. faster).}
\usage{
systemfit_unrelated(formula, df, sample_systemfit)
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Generate a systemfit object from a truly unrelated system of regressions (equivalent to systemfit vs. faster).
}
\examples{
TBC
}
