% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemfit_var.R
\name{systemfit_var}
\alias{systemfit_var}
\title{Generate a varcov matrix for a system of truly unrelated equations - to be used as part of a linearHypothesis test.}
\usage{
systemfit_var(group_var, df, clust = "ind", cluster_var = NULL,
  systemfit_obj = NULL)
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Generate a varcov matrix for a system of truly unrelated equations - to be used as part of a linearHypothesis test.
}
\examples{
TBC
}
