% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_fun.R
\name{t_fun}
\alias{t_fun}
\title{Perform a two-sided treatment-effect t-test (clustered/non-clustered).}
\usage{
t_fun(df, outcome_var, assignment_var, cluster_var = NULL)
}
\value{
p-value [numeric]
}
\description{
\
}
