% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_df.R
\name{table_df}
\alias{table_df}
\title{Convert a frequency table to a data.frame.}
\usage{
table_df(x, orientation = "vert", round_digit = 1)
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Convert a frequency table to a data.frame.
}
\examples{
TBC
}
