% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_mod.R
\name{table_mod}
\alias{table_mod}
\title{Ordered frequency table formatted as a data.table.}
\usage{
table_mod(x)
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Ordered frequency table formatted as a data.table.
}
\examples{
TBC
}
