% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dt_DATA.R
\docType{data}
\name{test_dt}
\alias{test_dt}
\title{Simulated cohort & feature set.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 10000 rows and 13 columns.}
\usage{
data(test_dt)
}
\description{
Cohort-style data.table to faciliate function/code testing and validation.
}
\details{
Maintainer: Clara Marquardt
}
\keyword{datasets}
