% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_legend_bottom.R
\name{theme_legend_bottom}
\alias{theme_legend_bottom}
\title{ggplot theme.}
\usage{
theme_legend_bottom(title_size = 0.5, text_size = 0.4, tick_size = 0.08,
  legend_width = 0.5, legend_height = 0.2, hjust_title = 0.5,
  font_gen = "URWHelvetica", col_gen = "grey50")
}
\description{
\
}
