% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_cat.R
\name{var_cat}
\alias{var_cat}
\title{Classify a variable in a dt according to a given classification scheme.}
\usage{
var_cat(dt, var_name, class_scheme)
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Classify a variable in a dt according to a given classification scheme.
}
\examples{
TBC
}
