% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_cat_gen.R
\name{var_cat_gen}
\alias{var_cat_gen}
\title{Generate classification schemes.}
\usage{
var_cat_gen(class_scheme_name)
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Generate classification schemes.
}
\examples{
TBC
}
