% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_overview.R
\name{var_overview}
\alias{var_overview}
\title{Generate a basic overview of a dataset.}
\usage{
var_overview(data, observation_unit = NA, alternative_id = NA,
  alternative_id_name = "/", data_name = NA)
}
\arguments{
\item{data}{data.table for which the overview is to be generated (data.table).}

\item{observation_unit}{Observation unit (character).}

\item{alternative_id}{name of column corresponding to the alternative_id_name (character).}

\item{alternative_id_name}{alternative observation unit, e.g. "patients" in a diagnosis file with multiple diagnoses per patient (character).}

\item{data_name}{file title (printed at the top of the overview) (character).}
}
\value{
Overview of dataset [data.table]
}
\description{
Dynamically generate an overview of a given dataset taking into account variable types.
}
\details{
Maintained by: Clara Marquardt
}
\examples{
summary_dt <- var_overview(dia, alternative_id="empi", alternative_id_name="patients",observation_unit="diagnoses", data_name="dia.rda" )
print(str(summary_dt))
}
