% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_append_csv.R
\name{write_append_csv}
\alias{write_append_csv}
\title{Paste multiple tables/character strings into a single csv file.}
\usage{
write_append_csv(element_list, file_path)
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Paste multiple tables/character strings into a single csv file.
}
\examples{
TBC
}
