% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xlsx_mod.R
\name{write_xlsx_mod}
\alias{write_xlsx_mod}
\title{Write toxslx sheet with a number of formatting options and safety checks.}
\usage{
write_xlsx_mod(dt, file, sheetName, lock = FALSE, password = "",
  unlock_col = NA, col_width = NA, alignment_var = "left", bold = FALSE,
  format_pattern = "")
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Write toxslx sheet with a number of formatting options and safety checks.
}
\examples{
TBC
}
