% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsMod.R
\name{writeCheckData}
\alias{writeCheckData}
\title{Write Check File as CSV}
\usage{
writeCheckData(dat, filename, msg, decor = TRUE, ...)
}
\arguments{
\item{dat}{a data.frame}

\item{filename}{name of output file}

\item{msg}{character text to display}

\item{decor}{logical value to indicate decoration before and after the msg}

\item{\dots}{additional arguments to pass to \code{\link{write.csv}}}
}
\description{
A check file contains IDs and data that need to be manually verified.
}
\keyword{internal}
