% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zocat_data.R
\docType{data}
\name{zocat}
\alias{zocat}
\title{Zocat-icd9 crosswalk.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 15071 rows and 7 columns.}
\usage{
data(zocat)
}
\description{
Crosswalk between icd9 diagnosis codes and zocat diagnosis categories.
}
\keyword{datasets}
