% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beautify_var_overview.R
\name{beautify_var_overview}
\alias{beautify_var_overview}
\title{: Format a given var_overview() outcome as a formatted two tab .xlsx sheet.}
\usage{
beautify_var_overview(var_overview_dt, data_definition_dt = NA, output_file,
  source = "", sheet_title = "Data Summary")
}
\arguments{
\item{var_overview_dt}{df returned by var_overview() (data.frame from var_overview)}

\item{data_definition_dt}{notes to be appended at the end of the overview, such              as variable definitions and other points. This *MUST* be a 2 COLUMN data.table (data.table)}

\item{output_file}{path and filename to output resulting Excel to (character)}

\item{source}{source of data / cohort name (to print in Excel) (character)}

\item{sheet_title}{Title of overview sheet (character)}
}
\description{
: Convert "var_overview" returned by Clara Marquardt's var_overview function into a production ready Excel file, with a separate tab for an overview of the data and a different one for a list of summary statistics from the data.
}
\examples{
\dontrun{
var_overview_dt <- var_overview(dem)
beautify_var_overview(var_overview_dt = var_overview_dt, output_file = "Summary.xlsx")
}
}
