% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_icd_list.R
\name{clean_icd_list}
\alias{clean_icd_list}
\title{Expand a list of icd9 code ranges.}
\usage{
clean_icd_list(range_table, code_col, name_col, decimal = TRUE,
  validate = FALSE)
}
\arguments{
\item{range_table}{data.table containing a column with icd9 code ranges and code range names (data.table).}

\item{code_col}{name of the column containing the icd9 code ranges (see the below example) - commented out rows (start with '#') are ignored (character).}

\item{name_col}{name of the column containing the code range names (character).}

\item{decimal}{whether or not the icd9 codes are in decimal format (logical - TRUE/FALSE) [default: TRUE].}

\item{validate}{whether or not to subset to 'defined' icd9 codes (see www.rdocumentation.org/packages/icd/versions/2.2/topics/icd_expand_range for more detail) (logical - TRUE/FALSE) [default: FALSE].}
}
\value{
data.table with the expanded icd9 code list and the associated code range names.
}
\description{
Expand a list of icd9 code ranges.
}
\examples{
orig_table <- copy(gagne_code) ## use the gagne_code sample data set
print(orig_table$code)         ## print the code range column
code_table  <- clean_icd_list(range_table=orig_table, code_col="code", name_col="condition", decimal=FALSE, validate=FALSE) 
print(code_table)
}
