% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseUtils.R
\name{collapse-internal}
\alias{collapse-internal}
\alias{qrbind}
\alias{nowarnnum}
\alias{freqNum}
\alias{parseFreq}
\alias{most}
\alias{pairDay}
\alias{mergeAdjacent}
\alias{borrowFreqDoseSeq}
\alias{borrowVal}
\alias{reOrder}
\alias{borrowWithinDoseSeq}
\alias{calcDailyDose}
\alias{daySeqDups}
\alias{rmDuplicates}
\alias{standardize_time}
\alias{convert_military}
\alias{convert_morn_night}
\alias{convert_ampm}
\alias{get_dur}
\title{Internal functions for collapse process}
\description{
These internal functions aid the main functions used in the collapse process (\code{collapse}, \code{makeDose}).
}
\details{
\code{qrbind}: fast version of rbind.data.frame

\code{nowarnnum}: converts a variable to numeric and suppresses warnings

\code{freqNum}: converts the frequency entity to numeric

\code{parseFreq}: standardizes the frequency entity

\code{most}: returns the mode

\code{pairDay}: calculates daily dose for dose sequences

\code{mergeAdjacent}: collapses adjacent rows that have a drug name change

\code{borrowFreqDoseSeq}: imputes a frequency of 'pm' if the preceding row has a frequency of 'am' or 'noon' and there is a 'pm' elsewhere in the note

\code{borrowVal}: borrow a unique value within a drug mention

\code{reOrder}: order rows

\code{borrowWithinDoseSeq}: borrow values within a dose sequence

\code{calcDailyDose}: calculates daily dose

\code{daySeqDups}: look for duplicate dose sequence 

\code{rmDuplicates}: removes redundant rows

\code{convert_ampm}: convert time expressions that explicitly use \sQuote{am} or \sQuote{pm} into 
HH:MM:SS format

\code{convert_morn_night}: if night/morning used, e.g. \sQuote{10 last night}, convert to am or pm

\code{convert_military}: convert military time into standardized HH:MM:SS format

\code{get_dur}: for duration expressions, e.g. \sQuote{12 hour level}, the numeric portion is extracted 
but not converted to a standardized time yet since it requires knowledge of lab time.

\code{standardize_time}: applies \code{convert_military}, \code{convert_morn_night}, \code{convert_ampm},  
and \code{get_dur} to create a string with standardized form of time in HH:MM:SS format
}
\keyword{internal}
