% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Collapse Dose Data}
\usage{
collapse(x, noteMetaData, naFreq = "most", ...)
}
\arguments{
\item{x}{data.frame}

\item{noteMetaData}{data.frame with columns filename, pid, date, note}

\item{naFreq}{Replacing missing frequencies with this value, or the most
common.}

\item{\dots}{drug names to split by}
}
\value{
A list containing two dataframes, one with the note level and one with the date level collapsed data.
}
\description{
Splits drug data and calls \code{makeDose} to collapse at the note and date level.
}
\details{
If different formulations of the drug (e.g., extended release) exist, they can be separated using a regular expression. This function will call \code{makeDose} on parsed and paired medication data to calculate dose intake and daily dose and remove redundancies at the note and date level.
}
