\docType{data}
\name{dd.baseline.small}
\alias{dd.baseline.small}
\title{dd.baseline.small}
\format{A data frame with 2000 observations on 55 variables.}
\usage{
dd.baseline.small
}
\description{
A smaller subset of baseline data example, dd.baseline. The dataset includes 55 
variables: subject identification number ('id'), drug exposure ('exposure'), 
3 demographic variables ('age', 'race', 'gender'), and 50 phenotypes.
}
\examples{
data(dataPheWASsmall)
}
\keyword{datasets}
