\docType{data}
\name{dd.small}
\alias{dd.small}
\title{dd.small}
\format{A data frame with 2000 observations on 55 variables.}
\usage{
dd.small
}
\description{
A smaller subset of outcome data example, 'dd'. The dataset includes 55 variables: subject identification number ('id'), drug exposure ('exposure'), 
3 demographic variables ('age', 'race', 'gender'), and 50 phenotypes.
}
\examples{
data(dataPheWASsmall)
}
\keyword{datasets}
