% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_data.R
\docType{data}
\name{dem}
\alias{dem}
\title{Simulated EHR data.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 400 rows and 8 columns.}
\usage{
data(dem)
}
\description{
Simulated EHR data (no PHI constraints) featuring dem data for ~500 patients
}
\keyword{datasets}
