% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dia_data.R
\docType{data}
\name{dia}
\alias{dia}
\title{Simulated EHR data.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 5893 rows and 9 columns.}
\usage{
data(dia)
}
\description{
Simulated EHR data (no PHI constraints) featuring dia data for ~500 patients
}
\keyword{datasets}
