% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed_data.R
\docType{data}
\name{ed}
\alias{ed}
\title{Simulated EHR data.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 869 rows and 11 columns.}
\usage{
data(ed)
}
\description{
Simulated EHR data (no PHI constraints) featuring ed data for ~500 patients
}
\keyword{datasets}
