\name{EHR-package}
\alias{EHR-package}
\alias{EHR}
\docType{package}
\title{
Electronic Health Record (EHR) Data Processing and Analysis Tool
}
\description{
Process and analyze Electronic Health Record (EHR) Data. Implement 
three statistical methods for Phenome Wide Association Study (PheWAS).
}
\details{
Contingency tables for many binary outcomes (e.g., phenotypes) and 
a binary covariate (e.g., exposure) can be efficiently generated by 
\code{\link{zeroOneTable}}, and three commonly used statistical methods 
to analyze data for PheWAS can be implement by \code{\link{analysisPheWAS}}.
}
\author{
Leena Choi <naturechoi@gmail.com> and Cole Beck <cole.beck@Vanderbilt.Edu>

Maintainer: Leena Choi <naturechoi@gmail.com>

}

\keyword{EHR}
\keyword{process}
\keyword{PheWAS}
