% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsExtractMed.R
\name{extractMed-internal}
\alias{extractMed-internal}
\alias{getNote}
\alias{getDose}
\title{Internal functions for Extract-Med module}
\usage{
getNote(note)
}
\description{
These internal functions aid the main function used in the Extract-Med module (\code{\link{extractMed}}).
}
\details{
\code{getNote}: Takes the file name and reads in the note as a character vector

\code{getDose}: Runs \code{\link[medExtractR]{medExtractR}} on a single clinical note (file) and labels 
the results output with file name
}
\keyword{internal}
