% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_logit.R
\name{feature_logit}
\alias{feature_logit}
\title{Generate, format and save a tabulation of the most important predictors
in a logit model (glm/glmnet/speedglm).}
\usage{
feature_logit(model, cluster_var_vector = NA, feat_lim = 200, output_path,
  add_dt = NULL)
}
\arguments{
\item{model}{}

\item{cluster_var_vector}{}

\item{output_path}{}
}
\description{

}
