% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pdf.R
\name{generate_pdf}
\alias{generate_pdf}
\title{Convert a list of ggplots into a (multi-page) PDF.}
\usage{
generate_pdf(plot_list, graph_cat = length(plot_list), ncol_plot, nrow_plot,
  file_path, file_name, orientation = "vertical", share_legend = FALSE,
  quiet = FALSE, height_plot = NULL, width_plot = NULL,
  height_overall = NULL, width_overall = NULL)
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Convert a list of ggplots into a (multi-page) PDF.
}
\examples{
TBC
}
