% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_race.R
\name{group_race}
\alias{group_race}
\title{: Classify given race as one of the conventional medical groups- white, black, hispanic, other.}
\usage{
group_race(data, race_col = "race", inplace = FALSE)
}
\arguments{
\item{data}{a data.table with the column containing races (data.table)}

\item{race_col}{a string specifying the name of column containing race info, "race" by default (character)}

\item{inplace}{if TRUE old race column is replaced with aggregated race infor, if FALSE (default), a new column called "race_group" is added (logical)}
}
\value{
data.table with the a new race_group column containing aggregated race / old column replaced with this column if inplace = TRUE (data.table)
}
\description{
Convert a column in a data.table containing nuanced, specific races into the aggregated set: {white, black, hispanic, other} that is more commonly found in US medical literature.
}
\examples{
mod_dem <- group_race(dem, inplace = TRUE)
unique(mod_dem[, race, ], by = "race")

grp_dem <- group_race(dem)
unique(grp_dem[, .(race, race_group), ], by = "race")
}
