% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_race.R
\name{group_race}
\alias{group_race}
\title{: Classify given race as one of the conventional medical groups- white, black, hispanic, other.}
\usage{
group_race(race_col)
}
\arguments{
\item{race_col}{data table column name pertaining to race. (object)}
}
\description{
Convert a column in a data.table containing nuanced, specific races into the aggregated set: {white, black, hispanic, other} that is more commonly found in US medical literature.
}
\examples{
dt <- copy(dem)
dt[, race := group_race(race)]

dt2 <- copy(dem)
dt2[, grouped_race := group_race(race)]
}
