% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastdoseUtils.R
\name{lastdose-internal}
\alias{lastdose-internal}
\alias{matchLastDose}
\alias{mergeLastDose}
\title{Internal functions for lastdose}
\description{
These internal functions incorporate lastdose.
}
\details{
\code{matchLastDose}: match lastdose data to built data

\code{mergeLastDose}: merge lastdose data into built data
}
\keyword{internal}
