% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_table.R
\name{list_table}
\alias{list_table}
\title{Convert list to table which, optionally, is appended to existing table.}
\usage{
list_table(list_name, var_name, argument = "space", merge = FALSE,
  old = NA)
}
\arguments{
\item{list_name}{List name [character]}

\item{var_name}{Name of variable which contains the header of the generated table (if merge==T 
and var_name in names(old)) -> the system time is appended to var_name to ensure 
column name uniqueness) [character]}

\item{argument}{List elements which equal this string are converted into empty rows in the final 
table (see \code{\link{list_space}})  [character]}

\item{merge}{Whether or not to merge the generated table with an existing [logical]}

\item{old}{Path to existing csv file with which generated table is to be merged
if merge==T [path]}
}
\value{
Generated table [data.table]
}
\description{
Convert list to table which, optionally, is appended to existing table.
}
\examples{
sample_list <- list()
sample_list$nfold <- 3
sample_list$test_prop <- 0.3
sample_list_dt <- list_table(sample_list, var_name="sample_run", merge=F )
}
