% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_or_install.R
\name{load_or_install}
\alias{load_or_install}
\title{Load or install R packages (public or local).}
\usage{
load_or_install(package_names, custom_lib_path = FALSE, custom_path = NA,
  verbose = FALSE, local_package = FALSE, local_package_path = NA)
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Load or install R packages (public or local).
}
\examples{
TBC
}
