% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_or_install_local.R
\name{load_or_install_local}
\alias{load_or_install_local}
\title{Load or install R packages (Local).}
\usage{
load_or_install_local(package_list, custom_lib_path = "", quiet = FALSE)
}
\arguments{
\item{package_list}{List of package paths which are to be installed/loaded (list - character).}

\item{custom_lib_path}{Custom library path (character) [Default: Default library path].}

\item{quiet}{Verbosity (logical - TRUE/FALSE) [Default: TRUE].}
}
\value{
List of packages which were succesfully installed/loaded.
}
\details{
Maintained by: Clara Marquardt
}
\examples{
\dontrun{
package <- list("[package path]", "[package path]") ## do NOT end in a /
load_or_install_local(package_list=package, custom_lib_path=paste0(getwd(), "/test/"), quiet=TRUE)
}
}
