% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_ci_clust.R
\name{mean_ci_clust}
\alias{mean_ci_clust}
\title{Calculate the clustered confidence interval for a sample mean or proportion.}
\usage{
mean_ci_clust(var, cluster_var, ci, return_sd = TRUE, digit = 2,
  min_lim = NA, max_lim = NA, type = "perc_prop")
}
\value{
Confidence interval (and optionally standard deviation) [character vector]
}
\description{
(1) Clustered CI for proportions based on: 
www.fao.org/wairdocs/ILRI/x5436E/x5436e07.htm
(2) Clustered CI for means based on: diffuseprior.wordpress.com/2012/06/15/\
standard-robust-and-clustered-standard-errors-computed-in-r/
}
\examples{
mean_ci_clust(as.integer(is.na(dem$date_of_death))*100,dem$empi, 
0.95, type="perc_prop")  # clustered prop CI
}
