% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_pdf.R
\name{merge_pdf}
\alias{merge_pdf}
\title{Merge a list of PDFs into a multi-paage PDF.}
\usage{
merge_pdf(input_file_list, output_file_name, quiet = FALSE)
}
\arguments{
\item{TBC}{}
}
\value{
TBC
}
\description{
Merge a list of PDFs into a multi-paage PDF.
}
\examples{
TBC
}
