% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_pdf.R
\name{merge_PDF}
\alias{merge_PDF}
\title{Merge a list of PDFs into a multi-page PDF.}
\usage{
merge_PDF(pdf_list, output_file)
}
\arguments{
\item{pdf_list}{List of file paths to the individual PDF files which are to be merged (list - character).}

\item{output_file}{Path/filename at which the merged PDF file is to be saved (character).}
}
\value{
Merged PDF saved to specified path.
}
\description{
Given a list of paths to locally stored PDF files - merge the PDFs (in the given order) and save the newly created PDF to a specified location.
}
\details{
Maintained by: Clara Marquardt
}
\examples{
\dontrun{
pdf_list <- list("[path.pdf]", "[path.pdf]") 
merge_PDF(pdf_list=pdf_list, output_file="merge_pdf_test.pdf")
}
}
