% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsMod.R
\name{mod-internal}
\alias{mod-internal}
\alias{colflow_mod}
\alias{findMedFlowRow_mod}
\alias{flowData_mod}
\alias{concData_mod}
\alias{infusionData_mod}
\alias{resolveDoseDups_mod}
\alias{processErx}
\alias{processErxAddl}
\alias{validateColumns}
\alias{read}
\title{Internal functions for EHR modules}
\description{
code{colflow_mod}: build a comparison of potentially duplicate flow data
}
\details{
code{findMedFlowRow_mod}: find row numbers for matching ID and time

code{flowData_mod}: correct flow data

code{concData_mod}: correct concentration data

code{infusionData_mod}: combine flow and MAR data; standardize hourly rate,
which may or may not include weight

code{resolveDoseDups_mod}: correct duplicated infusion and bolus doses

code{processErx}: prepare e-prescription data

code{processErxAddl}: additional processing, checking for connected observations

code{validateColumns}: check columns are in place

code{read}: generic read function
}
\keyword{internal}
