% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsPk.R
\name{pk-internal}
\alias{pk-internal}
\alias{takeClosest}
\alias{takeClosestTime}
\alias{daysDiff}
\alias{minDiff}
\alias{fixDates}
\alias{combine}
\alias{updateInterval_mod}
\alias{fix_serum_time}
\alias{merge_by_time}
\alias{merge_inf_bolus}
\alias{addZeroDose}
\alias{pkdata}
\alias{parse_dates}
\alias{guessDateFormat}
\title{Internal functions for pk analysis}
\description{
code{takeClosest}: take value at nearest time
}
\details{
code{takeClosestTime}: find nearest time

code{daysDiff}: number of days between two dates

code{minDiff}: difference (in minutes) of time vector

code{fixDates}: add leading zeroes to date

code{combine}: rbind two data sets (GIVE DIFFERENT NAME)

code{updateInterval_mod}: replace maxint with difference between surgery time
and infusion time (if smaller than maxint)

code{fix_serum_time}: create time variable

code{merge_by_time}: impute in x by closest time in y

code{merge_inf_bolus}: combine infusion and bolus into one hourly dose file

code{addZeroDose}: add rows of zero dose, after a gap

code{pkdata}
}
\keyword{internal}
