% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processLastDose.R
\name{processLastDose}
\alias{processLastDose}
\title{Process and standardize extracted last dose times}
\usage{
processLastDose(mxrData, noteMetaData, labData)
}
\arguments{
\item{mxrData}{data.frame containing output from the \code{medExtractR} system}

\item{noteMetaData}{data.frame with meta data (\code{pid} (patient ID) and \code{date}) for 
the file names contained within \code{mxrData}}

\item{labData}{data.frame that contains lab dates and times associated with the file names 
within \code{mxrData}. Must contain columns \code{pid} and \code{date}, as well as \code{labtime}. 
The \code{date} column must be in the same format as \code{date} in \code{noteMetaData}, and 
\code{labtime} must be a POSIXct}
}
\value{
data.frame with identifying information (e.g., filename, etc) as well as processed 
and standardized last dose times as a POSIXct column
}
\description{
This function takes last dose times extracted using the \code{medExtractR} system and 
processes the times into standardized datetime objects using recorded lab data.
}
\examples{
\donttest{
# add data before finalizing example
processLastDose(mxrData = df, noteMetaData = meta_data, labData = tac_lab)
}
}
