% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsId.R
\name{pullFakeId}
\alias{pullFakeId}
\title{Pull Fake/Mod ID}
\usage{
pullFakeId(
  dat,
  xwalk,
  firstCols = NULL,
  orderBy = NULL,
  uniq.id = "subject_uid"
)
}
\arguments{
\item{dat}{a data.frame}

\item{xwalk}{a data.frame, providing linkage for each ID}

\item{firstCols}{name of columns to put at front of output data set}

\item{orderBy}{name of columns used to reorder data set}

\item{uniq.id}{character sting indicating subject-level id variable (default is "subject_uid")}
}
\value{
The modified data.frame
}
\description{
Replace IDs with de-identified version pulled from a crosswalk.
}
