% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_mult_structure.R
\docType{data}
\name{return_mult}
\alias{return_mult}
\title{Store multiple arguments returned from function as separate objects - structure.}
\format{An object of class \code{result} of length 1.}
\usage{
return_mult
}
\value{
TBC
}
\description{
Store multiple arguments returned from function as separate objects - structure.
}
\examples{
\dontrun{
within function_test: return(list(obj_1,obj_2, obj_3))
outside of function: return_mult[obj_1, obj_2, obj_3] <- function_test()
}
}
\keyword{datasets}
