% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modLabs.R
\name{run_Labs}
\alias{run_Labs}
\title{Run Lab Data}
\usage{
run_Labs(lab.path, lab.select, lab.mod.list)
}
\arguments{
\item{lab.path}{filename of a lab file (CSV, RData, RDS), or data.frame}

\item{lab.select}{columns to select}

\item{lab.mod.list}{list of expressions giving modifications to make;
passed to \code{\link{dataTransformation}}}
}
\value{
lab data set
}
\description{
This module will load and modify laboratory data.
}
\details{
See EHR Vignette for Structured Data.
}
\examples{
lab_data <- data.frame(mod_id=rep(1:3,each=3),
                       date=rep(c("01/12/17","05/05/18","11/28/16"),each=3),
                       time=rep(c("1:30","2:30","3:30"),3),
                       creat=rnorm(9,0.5,0.05))
run_Labs(lab_data, lab.mod.list=list(log_creat=expression(log(creat))))

}
