% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_overview.R
\name{var_overview}
\alias{var_overview}
\title{Generate an overview of a dataset.}
\usage{
var_overview(DT, alt_id = NA, alt_id_name = "", var_def_merge = FALSE,
  var_def_file = NA, observation_unit = NA, file_title = NA)
}
\arguments{
\item{DT}{Name of DT for which summary stats are to be generated}

\item{alt_id, alt_id}{name Name of id variable on the basis of which to generate the 
observation count (set to NA if no relevant identfier exists)}

\item{var_def_merge}{Whether to merge in csv file with variable definitions}

\item{var_def_file}{Path of csv file with variable definitions merged in if var_def_merge==TRUE}

\item{observation_unit}{Unit of observation (e.g."diagnoses","encounters")}

\item{file_title}{Dataset title to printed at the top of the summary statistics}
}
\value{
Overview of dataset [data.table]
}
\description{
Generate an overview of a dataset.
}
\examples{
summary_stat <- var_overview(dia, alt_id="empi", alt_id_name="patient",observation_unit="diagnoses", 
file_title="dia.rda" )
}
